+++ CHANGELOG.md
@@ -1,8 +1,6 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
-
-## Unreleased
+## 0.2.0 - 2026-01-10
 
 ### Added